piTest: Fix return value on failure of piControlOpen()
If the call to piControlOpen() fails in piControlRead(),
piControlWrite(), piControlGetBitValue() or piControlSetBitValue(),
-1 is returned which is decoded by getReadError() / getWriteError() to
"Cannot connect to control process".

In all other cases, it seems more useful to return the actual reason
why piControlOpen() failed, instead of returning the magic number -18
(-EXDEV = Cross-device link) or -1 (-EPERM = Operation not permitted).
The reason is stored in PiControlHandle_g.

Signed-off-by: Lukas Wunner <lukas@wunner.de>