piTest: Handle reading/writing fewer bytes than requested
read(2) guarantees to read the number of bytes requested only "if the
descriptor references a normal file that has that many bytes left before
the end-of-file", but in particular not for a device file.  (Quote taken
from the manpage.)

Likewise write(2) does not guarantee to actually write the number of
bytes requested.

Treat such cases as an error instead of silently ignoring them.

Signed-off-by: Lukas Wunner <lukas@wunner.de>