Improve stability of communication with PiGates
The communication with PiGate modules was not very reliable. The state
machine was changed to improve the stability. The change works with all
Gateway, old releases and new release that use the same source code.
This code contains a lot of debug messages, they must be removed in the
final release.