piControl: Introduce cycletimer
The traditional jiffies-based sleep routines such as schedule_timeout()
are too imprecise, so an hrtimer-based mechanism is needed.

In principle, hrtimer_nanosleep() does what we want but for one, it's
not exported, and for another it generates an hrtimer on the stack on
every invocation, which is inefficient.

Introduce struct cycletimer, which also takes care of detecting and
reporting missed cycles.

Signed-off-by: Lukas Wunner <lukas@wunner.de>