piControl: Support reset ioctl on RevPi Compact
The only thing that's configurable on the RevPi Compact at the moment
is the ain channels.  On reset, the ain poll thread is stopped (if it
is running at all) and restarted.  It automatically determines which
channels to poll from now on.  Channels which are no longer polled are
represented in the process image with their last known value.  Channels
which remain enabled are visible in the process image without glitches.

Signed-off-by: Lukas Wunner <lukas@wunner.de>