piControl: Future-proof sched_param initialization
Currently struct sched_param is not initialized and later on only the
sched_priority element is set.  Right now that is fine as there are no
other elements in the struct, hence no risk of using uninitialized data,
but that might change in the future so initialize the entire struct.

Signed-off-by: Lukas Wunner <lukas@wunner.de>