Add support for MasterGateModule and reworked LED control
If a Master Gate Module has to be used with the RevPi a Host application
is used for configuration. This configuration is transferred as binary
data to Master Module using the serial line on the PiBridge. Three new
ioctl-calls ar used to stop the normal I/O communication, transfer the
config data and restart the I/O communication. To use thsi a deamon must
be started on the RevPi to communicate with the configuration
application with a TCP/IP socket and call the ioctls.