Add watchdog for output
If an application set output cyclically the new watchdog can be used to
set all output to 0 if the application crashes.
The watchdog must be activated with a call of the ioctl
KB_SET_OUTPUT_WATCHDOG. The argument is a pointer to an unsigned int
variable with the watchdog duration in ms. The watchdog timer is started
each time the function write is called. The application must call
write() in shorter periods than the watchdog duration. When write is not
called anymore, all output are set to 0. Also when the application is
terminated or crashed all output are set to 0. The watchdog can be
deactivated by setting the duration to 0.