piControl: Add module softdeps for RevPi Compact
The SPI-attached I/O chips need to be accessible for revpi_compact_init()
to succeed.  However their modules are not necessarily loaded before
piControl because it doesn't reference any of their symbols.  The only
exception is max3191x_get_status().  Specify all other modules as a
softdep to ensure their presence.

On the RevPi Core we depend on presence of the spi0 driver.  So far
we're satisfying that by explicitly loading the module from within
piControl.  Use a softdep here as well for consistency.

Signed-off-by: Lukas Wunner <lukas@wunner.de>