piControl: Set RT priority for required kthreads on init
On the Revpi Core, the spi0 kthread needs to be assigned a realtime
priority to timely service transmissions to attached RevPi Gateways.

The original wheezy image assigned the realtime priority in an init
script which also loaded the piControl module.

With jessie, we wanted to just list piControl in /etc/modules without
requiring an additional init script or systemd service, so we defined
the realtime priority in the devicetree and assigned it with a small
kernel patch.

However that approach doesn't scale to the requirements of new products:
On the RevPi Compact, we now need to assign a realtime priority to the
spi2 and ktimersoftd kthreads and we wouldn't want to write a kernel
patch for each of them.  Instead, let's migrate to a new concept wherein
the realtime priorities are defined per-product in piControl.

Currently there are no provisions to revert the kthread priority on
unload of piControl, it remains assigned indefinitely.

Signed-off-by: Lukas Wunner <lukas@wunner.de>