piControl: Use native network stack for RevPi Gate communication
Previously the RevPi Gate communication was implemented on top of the
raw spi_device and piControl duplicated the functionality of ks8851.c
in the kernel as well as all link layer handling.

Reimplement the RevPi Gate protocol on top of the Linux-native link
layer handling by registering a protocol handler, thereby deduplicating
code, allowing for easier porting to other Ethernet chips and allowing
for simultaneous use of RevPi Gates with SPI-attached RevPi Connect
modules (such as RevPi Con CAN).

The KUNBUS header files remain mostly unmodified and can continue to be
shared with the firmware running on the RevPi Gates.

Signed-off-by: Lukas Wunner <lukas@wunner.de>