piControl: Start UART kthread last on module load
On the Revpi Core and Connect, after starting the UART and I/O kthreads,
the RT priority of the spi0 kthread is adjusted.  If the spi0 kthread
hasn't been spawned yet, piControl tries to tear down the UART kthread
which consistently deadlocks due to its design.  To the user this is
visible as a boot hang.

Avoid the boot hang by adjusting the spi0 priority before spawning the
UART and I/O kthreads.  piControl will still fail to load, but at least
the system is accessible to the user.

Signed-off-by: Lukas Wunner <lukas@wunner.de>