piControl: Delay initialization until after driver probing
When piControl initializes on the Revpi Core or Connect, the RT priority
of the spi0 kthread is adjusted.  If the spi0 kthread hasn't been
spawned yet because the spi0 driver hasn't finished probing,
initialization of piControl fails.

Avoid by explicitly enumerating the spi0 driver as a soft dependency
and waiting for all devices to finish probing before initializing
piControl.

Signed-off-by: Lukas Wunner <lukas@wunner.de>