piControl: Fix module init error paths
Upon initialization, piControl keeps track of the steps it has carried
out by incrementing the init_step variable.  The error paths and module
remove path then tear down data structures by checking how far
initialization got.  The idea is to use the same code in the error paths
and for module removal.  However that scheme is somewhat unusual and
doesn't really square with kernel coding style, which generally uses
gotos for error handling.  Moreover there are multiple bugs in the
teardown code which are difficult to fix with the current scheme.

Migrate to gotos and fix those bugs.

Signed-off-by: Lukas Wunner <lukas@wunner.de>