piControl: Silence unsupported ioctl errors
Patrick reports that simply opening the piControl chardev in Python
results in an "Invalid Ioctl" message.  According to Sven Sager, Python
issues the ioctl() to test whether it is talking to a tty.

Opening the chardev repeatedly in this way fills up the logs, which is
undesirable.  Avoid by toning down the message to debug severity.

Link: https://revolution.kunbus.de/forum/viewtopic.php?f=6&t=1041
Signed-off-by: Lukas Wunner <lukas@wunner.de>