piControl: compact: Fix stale CPU freq and temp if ain is disabled
Since commit 93a76bbc4a9c ("Stop I/O communication for simulation"), the
revpi_compact_poll_ain() kthread is not only responsible for reading the
analog inputs, but also for reading the CPU frequency and temperature.

However if all analog inputs are disabled, the kthread is put to sleep
and the CPU frequency and temperature are thus not read.  Their values
in the process image become stale.

Fix by waking the kthread once per second even if all analog inputs are
disabled.

Signed-off-by: Lukas Wunner <lukas@wunner.de>