piControl: compact: Fix dev_remove_pack() error on driver reset
Since commit 2de8667a3407 ("piControl: Use native network stack for
RevPi Gate communication"), issuing a driver reset on the RevPi Compact
leads to an error such as:

    dev_remove_pack: 2429d366 not found

That's because PiBridgeMaster_Stop() is called upon driver reset even
though the Compact lacks a PiBridge.  Constrain the call to the Core
and Connect.

Signed-off-by: Lukas Wunner <lukas@wunner.de>