piControl: compact: Rework cycletimer
cycletimer_change() is invoked upon initialization of both the ain and
the io thread on the Compact, as well as upon reset of the ain thread.

Because of the way it currently works, it frequently causes a warning on
the next invocation of cycletimer_sleep() because the hrtimer either
hasn't expired yet or a cycle was missed.

Rework the function to commence a new cycle upon its invocation, thus
allowing for a full cycle time until cycletimer_sleep() must be called
to avoid any warnings.

Signed-off-by: Lukas Wunner <lukas@wunner.de>