piControl: add missing lock
As a comment implies the author of the code did not see the requirement of
locking the copy of a byte value in case of the KB_GET_VALUE ioctl.
Unfortunately this assumption is wrong:
In case of two threads A (writer) and B (reader) that access the same data
we have to ensure memory coherency. Thus for thread A we must ensure that
the written value actually is written to its RAM destination and not held
in cache. For thread B we must ensure that the value is actually read from
RAM and not from cache. Since the mutex lock()/unlock() operations imply
the required memory barriers both can be achieved by using mutex.

The writer already uses a mutex to access the data. Use this for the
reader, too.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>