piControl: compact: Hold ref on iio_dev until channel acquisition
revpi_compact_init() acquires a first ref on the ADC and DAC IIO devices
with bus_find_device().  The ref is then released before additional refs
are acquired with iio_channel_get_all() and iio_channel_get().  This
introduces a race wherein the iio_dev may disappear between release of
the first ref and acquisition of the additional refs.  Fix by holding
the first ref until after the calls to iio_channel_get_all() and
iio_channel_get().

Note that holding a ref on the device does not prevent unloading the
module backing the IIO driver.  Rather, the IIO accessors return an
error:
https://patchwork.kernel.org/patch/10332711/

I've verified that accessing the ADC and DAC IIO devices from piControl
after unloading the driver modules does not throw any errors and that
AIn_Status and AOut_Status are properly set to 0x80 when accessing the
IIO devices.

Reported-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>