piControl: compact: Fix iio_channel leak on init
revpi_compact_init() invokes iio_channel_get() for the two DAC IIO
channels and bails out on error.  The code works correctly if the
invocation fails for both channels.  But if it fails for a single
channel, the other one is leaked.  Fix by handling errors
individually for each channel and unwinding the first invocation
if the second failed.

Reported-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>
Signed-off-by: Lukas Wunner <lukas@wunner.de>