piControl: Protect access to UART receive buffer
piIOComm.c uses a cyclic buffer to store bytes received from the UART,
but lacks a lock to protect access to the head and tail pointers.
The recvLenSem, which protects access to i16uRecvLen_s, happens to be
taken by most code paths which access the head and tail pointers, but
not all of them.

Acquire it in the remaining code paths to prevent corruption of the head
and tail pointers through concurrent access.

Signed-off-by: Lukas Wunner <lukas@wunner.de>