piControl:compact: add missing locks in ain poll thread
At various places in the ain poller thread values are assigned to the
driver structure which is copied in the dio poller thread. During the copy
a lock is held in the dio thread, while the value assignment in the ain
thread is done without any locks.
This may lead to the assigned values not being visible immediately (or not
being visible at all) in the dio thread (e.g if both threads are running on
different CPUs and the CPU the ain thread is running on, keeps the values
in its cache).
Fix this by using the proper lock also within the poller thread.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>