piControl: Support more LEDs for event trigger
Function revpi_led_trigger_event() uses two masks (one for the previous and
one for the current active LEDs) to decide for which LED to trigger an
event. Since both masks are only 8 bit wide, only 8 LED states can be
distinguished. This is not sufficient for the REVPI Flat. So change the
function parameters to take 16Bit wide masks instead of the 8 bit wide
ones.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>