piControl:flat: Specify module dependencies
On the flat the module piControl depends on ad5446 which it uses for analog
output. However starting piControl on bootup may fail due to ad5446 not
being present at this time.
The reason is that there is that programs like modprobe do not know about
this module dependeny.
So give modprobe a hint by adding ad5446 to the list of modules specified
in MODULE_SOFTDEP.

Reported-by: Simon Han <z.han@kunbus.de>
Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>