piControl:flat: Use global struct to store configuration
Remove the structure that holds configuration data from struct revpi_flat
and instead use a static structure for this purpose.
This is in preparation of passing the configuration data from the parser to
the flat implementation (the data will be passed even before struct revpi_flat
has been allocated).

By creating the new structure also get rid of the "offset" configuration
setting since it has not been used anyway.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>