piControl:flat: apply default configuration
Pass the parsed configuration data and store it in the global
configuration data struct. Copy the stored data to the user data segment in
the reset function and call the reset function at module initialization.

This ensures that the default settings are applied in the worker threads
for both module startup and module reset.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>