piControl:compact: remove priority change of ktimersoftd thread
The ktimersoftd thread does not exist any more in newer kernels. So
remove the code to set its priority.
Note: using piControl not leads do a lot of messages like "piControl i/o:
missed X cycles". This may be caused by the removal of the priority
settings and has to be investigated further.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>