revpi: core: register platform driver for revpi core device
Make use of the new "pibridge" device tree entry and register a platform
driver that is compatible to this node. Move the device initialization/
deinitialization to the drivers probe/remove functions.
With this we dont have to choose the GPIO controller by name but get
the sniff pins out of the pibridge entry.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>