Use hardware interrupt timer instead in soft interrupt
The HRTIMER_MODE_ABS uses a softirq timer. This means the hrtimer will
wakup the softirqd and the softirqd will wakeup our thread. This causes
additional delays and jitter. We tried to combat this with increasing
the softirq priority.
The better solution is to wake our threads directly from hardware timer
interrupt. Thus using HRTIMER_MODE_ABS_HARD will do this for us.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>