compact: Shrink critical section (cpu temp/freq)
Calling thermal_zone_get_temp() and bcm2835_cpufreq_get_clock() under
the piDev_g.lockPI is causing large latencies in the io thread. In the
ain thread the lock is held while the cpu frequency and temperature is
read. For this calls in the GPU firmware are used. So the lock is held
for an extensive time. The io threade needs to wait on the lock in the
flip_process_image() macro.

The solution is simple. Read the cpu freq. and temp. without the lock
and update the process image afterwards.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>