flat: Shrink critical region in the ain thread
We read the CPU temp. and freq. in the AIN thread. We acquire the lock
before we read the CPU temp. and freq. It is not needed to hold this
lock to do so. And as this might take a lot of time we might block the
DOUT thread.

Reading the CPU temp. and freq. outside the lock and hold the lock only
to write the values to the process image.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>