piControl: compact: Avoid starving USB traffic
On kernel 5.10, piControl's high-priority tasks may starve USB traffic,
with keyboard and network URBs seemingly timing out.  The issue is
confined to dwc_otg and cannot be reproduced with dwc2. However, lack of
the FIQ optimization makes switching to dwc2 unattractive.

It turns out that raising the RT priority of dwc_otg's IRQ threads above
spi2 and piControl resolves the issue, at the expense of occasional
lost or premature cycles.  USB traffic is handled in softirq (tasklet)
context and it appears on kernel 5.10 that's done on exit from the
dwc_otg IRQ threads, whereas on 4.19 is was done by the ktimersoftd
threads.

The name of the dwc_otg IRQ threads contains the IRQ number, which is
variable.  In addition, there are two threads with identical name due
to TASK_COMM_LEN truncation.  Amend set_kthread_prios() to cope with
these two particularities.

Signed-off-by: Lukas Wunner <lukas@wunner.de>