piControl: compact: Lower priorities of kernel threads
For compact the piControl module raises the priorities for the SPI (spi2)
thread to a realtime priority of 61 and the priorities of the ain and i/o
thread to 56 and 58 repectively.
As it turned this leads to massive perfomance issues on the USB ethernet
device. The reason seems to be that the USB worker threads (irq/X-dwc_otg)
with the lower priority of 50 are starving now from time to time.

To fix this make sure that the spi2, ain and i/o threads are not higher
prioritized than the USB worker threads.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>