piControl: support connect se compatible
At module initialization the compatible string in the device tree is
checked to determine the device type. However there is no check for the
new compatible "kunbus,revpi-connect-se" yet, so that a Connect SE is not
recognized as connect, but as a core (which is the default type).

Fix this by checking against the new Connect SE type.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>