piControl: use serdev based revpi-comm device for RS485 communication
There is a serdev implementation for the RS485 piBridge-communication
available. Use that instead of establishing an own serial connection in
piControl. For now we only replace the low-level functions with
pibridge_send() and pibridge_recv_timeout() to keep code changes minimal.
An exception is the use of pibridge_req_gate_tmt() which covers a hole
transaction (request and reception of data with consecutive check of
received data; in a later step only high-level functions that cover
complete transactions should be used).

Clean up the code and remove everything that is not needed any longer,
especially the UART thread and the related data structures which are now
replaced by the functionality implemented in the serdev module.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>