piControl: ensure setting I/O processing state is threadsafe
The piControl kernel modules uses a global bool variable “piDev_g.stopIO”
which can be set by means of ioctls to steer the input/output value
processing: if the variable is set to “false” input and output are
processed as usual, while a value of “true” results in changing of
input/output values being supressed (i.e. not passed from/to userspace).

The problem is that the variable is not threadsafe. This means that if it
is changed by a userspace process the set value may or may not be visible
to the processor that is currently processing the input/output (loop). In
the worst case this may result in setting of the variable having no effect.

So ensure the data integrity by using the threadsafe bitoperations provided
by the kernel. For this turn the boolean value into a unsigned long and
define the bit to determine if I/O processing is enabled or not.

By doing this also correct a related comment in the source code.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>