piControl: Simplify DIO PWM comparison
When sending a cyclic request to a DIO, PWM values are only included if
they have changed since the last cycle.

The search for changed PWM values is overly complicated:  For reasons
that pass understanding, the PWM values are searched last-to-first.

Also, the digital output values are compared in addition to PWM values.
That doesn't make any sense as digital output values are *always*
transmitted, on every cycle.

Simplify the comparison by using memcmp() and ignoring digital output
values.

Signed-off-by: Lukas Wunner <lukas@wunner.de>