piControl: use pibridge_req_io() to exchange AIO data
In piAIOComm_sendCyclicTelegram() functions piIoComm_send() and
piIoComm_recv() are used to exchange I/O data with the AIO module. The
complete send-receive logic is also implemented in pibridge_req_io(). So
use this function instead to avoid code duplicity.

By doing this refactor the code to be closer to the linux kernel coding
style.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>