piControl: Migrate MIO config communication to pibridge library
Use the pibridge_req_io() helper to construct and send a request and
receive and validate the response instead of open coding the same.

Store only the config payloads instead of the full-fledged telegrams
in mio_list[].  That avoids the need for revpi_mio_addr_chk() because
inconsistencies between the telegram headers are no longer possible.

Add an "addr" member to each mio_list[] array element to discover the
payloads corresponding to a specific module.  (Previously the address
in the telegram header was used instead.)

Signed-off-by: Lukas Wunner <lukas@wunner.de>