piControl: ensure setting telegram pending state is threadsafe
The piControl kernel modules uses a global bool variable
“piCore_g.pendingUserTel” which can be set by means of ioctls to steer the
sending of telegrams. The IO-Thread then checks the variable then in each
loop and sends a telegram if it is set. It also clears the variable
thereafter.

The problem is that the IO-Thread does not use any locks when reading or
clearing the variable. So it is not guaranteed, that the value is preceived
the same by the IO-Thread as it has been set by the thread executing the
ioctl.

Ensure data integrity by using the threadsafe bitoperations provided by
the kernel. Instead of a boolean variable use a flag that is set or cleared
by the concurrent running threads.

Also make the code a bit simpler by not only setting but also clearing the
variable in the thread executing the ioctl.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>