Connect 4: Add digital output
The RevPi Connect 4 has one digital output which the user wants to
controll via the process image. Contrary to the existing RevPi Connect,
where the output is controlled through the RevPiLED byte, this is done
via the new output byte (union with old led byte). In order to make this
work, the value change logic from the led byte / word has been
duplicated.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>