Don't update the base module
A RevPi base module has no firmware, which can be updated via piControl.
Thus the base device (address 0) needs to be skipped in FWU_update().

Return EOPNOSUPP if a base module the FWU_update() is detected and ensure, that
ioctl reports success.

Customer were reporting this "issue" in the past and might think that something
is missing or broken on their device.

See https://revolutionpi.com/forum/viewtopic.php?t=4142 for example.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>