Move check for virtual device into FWU_update
The firmware update checks wether the module requested to update is a hardware
module or not. If the firmware update ioctl is called for a virtual module, the
current implementation will exit with EFAULT, which is wrong. The return code
should be EOPNOTSUPP, as is was implemented for the base device.

Fix this and move the check into FWU_update() and return EOPNOTSUPP if a
virtual module is detected.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>