FWU: Return error code to user space
The actual error code of the firmware update is never returned to the caller
application in user space and only logged into kernel log. The return value of
the ioctl is in case of error always EFAULT.

Change this to the real return value of FWU_update() and exit the update loop,
when a update failed.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>