piControl: Add support for RO module
Implement support for the RevPi Relay (RO) Module. The four relays are
represented as bits in a 8-bit bitfield in the process image and can be
switched on/off by setting/clearing the concerning bit. Each closed relay
is indicated by the light up of an individual green LED on the RO modules.

The same bits in another (status) byte represent the status of the relays.

With the initial configuration also an optional threshold can be set for
each relay. The threshold is used by the RO module to warn about relay
wearout:
If the number of times a relay has been switched equals or exceeds the
threshold, a warning is indicated by a set bit in the status byte.
Furthermore blinking red and green LEDs show this state in case of a closed
relay while a red light shows it in case of an open relay.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>