piControl: Add support for RO module
Implement support for the RevPi Relay (RO) Module. The four relays are
represented as a 8-byte bitfield in the process image (with one additional
byte reserved for future use) and can be switched on/off by
setting/clearing the concerning byte. A closed relay is indicated by a
green led on the RO modules. One further byte represents the status for
each relay.
With the initial configuration also an optional threshold can be set for
the relays. The threshold is used by the RO module to warn about relay
wearout:
If the number of times a relay has been switched exceeds the threshold, a
warning is indicated by a set bit in the status byte and by switching on
a red LED on the RO module.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>