piControl: fix telegram communication
Currently the telegram communication always returns an error. This is
because instead of the exact number of expected response bytes we pass the
max value for the response to function pibridge_req_io(). This function
however cannot be used for such a case since it expects to know the exact
number of bytes to receive beforehand.

Fix this situation by using the low level pibridge API (namely a
combination of pibridge_req_send_io() and pibridge_recv()) instead.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>