piControl: Avoid leaking response of previous telegram to user space
If a first user space application receives a large response to a user
telegram and a second application then receives a shorter response to
a user telegram, the last couple of bytes of the first response may be
leaked to the second application.

Avoid by zeroing the response buffer before reception.

Signed-off-by: Lukas Wunner <lukas@wunner.de>