piControl: Fix IOPROTOCOL_MAXDATA_LENGTH off-by-one errors
The bitLength field in UIoProtocolHeader is 5 bits long and thus allows
for a payload length of 0 .. 31 bytes.  Hence IOPROTOCOL_MAXDATA_LENGTH
must be defined to 31 bytes, not 32 bytes.

pibridge_req_io() expects the size of the response buffer to not include
the CRC.  Hence the maximum size is 31 bytes, not 32 bytes (which is
the size of the ai8uData field of SIOGeneric).

Signed-off-by: Lukas Wunner <lukas@wunner.de>