piControl: use packet attribute instead of header inclusion
The header files COMP_packBegin.h and COMP_packEnd.h are included before
and after the declaration of a structure that ought to be packed.
While COMP_packBegin.h is nowadays an empty file, the inclusion of
COMP_packEnd.h adds the "packed" attribute to the structure.
Remove both files and add the attribute explicitly to the concerning
structs.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>