piControl: fix structure size for 64 bit machines
Structure StrFPGAHeadData is supposed to be 8 bytes in size. However it
consists of two unsigned long variables which results in 8 bytes size on 32
bit machines but in 16 bytes on 64 bit machines. Avoid this by using 4-byte
fixed-sized datatypes like INT32U instead of unsigned long for the variable
data types.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>