piControl: Fix fwuWrite() length check
fwuWrite() sends one segment of a firmware update to a module by
encapsulating it in a gateway protocol packet.

The gateway protocol allows for a payload size of 255 bytes, of which
4 bytes are occupied by the offset into the firmware file.

fwuWrite() does perform a length check but inexplicably allows a payload
size of 300 bytes, not 255 bytes.

Luckily, the sole caller of fwuWrite(), PiBridgeMaster_Run(), performs a
length check of its own and constrains the segment size to 250 bytes
(MAX_FWU_DATA_SIZE).

For correctness, check for 255 bytes instead of 300 bytes.

While at it, change a nonsensical "<= 0" check to "== 0" as the length
is unsigned.

Signed-off-by: Lukas Wunner <lukas@wunner.de>
Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>