piControl: flat: Do not request wrong GPIO for top button on flat S
Loading piControl at boot time breaks the communication on SPI 5 (i.e. the
TPM is not accessbile). The reason is the attempt to configure the top
button GPIO when piControl is started:
While old Flat revisions use GPIO 13 for the top button this GPIO is used
for the MISO line of SPI 5 on the Flat S 2022. On Flat S 2022 however GPIO
23 is used for the top button.

Fix this issue by implementing a check for the FLAT variant and chosing the
proper GPIO for the top button configuration.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>