piControl: mio: clear last output values at reset
The mio code stores the last set analog output values and then compares it
with the values currently set in the process image. If a difference is
detected the number of changed channels are extracted and then only for
these channels the new values are sent to the MIO.

However currently the stored values are not cleared when piControl is reset
via KB_RESET ioctl. This results in the concerning analog output channels
not being set if the last stored values equal the default values that are
read at module reset and then stored in the process image.

Fix this by clearing the stored values and thus make sure that default
values for the analog output channels are initially sent to the MIO.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>