Replace bcm2835_cpufreq_get_clock() with cpufreq_quick_get()
Calling bcm2835_cpufreq_get_clock() get the cpu frequency from through
the rpi firmware. Which needs to send a message obver the mailbox system
to the gpu. This can create huge delays (more the 100ms).

Instead use cpufreq_quick_get() which needs the cpu# as argument. As all
cores of the BCM283x, BCM2711 and BCM2712 have the same clock, we can
use the first cpu core an all devices.

The documentation of cpufreq_quick_get() says:
> This is the last known freq, without actually getting it from the driver.

Which indicates, that the value might not be up to date on every call.
This should not be an issue as we don't do anything elaborate with the
clock frequenzy. It has only informational value for the user. It is the
same  value as can be found in scaling_cur_freq in the sysfs.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>