piControl: bump module version to 2.0.0
piControl has seen a lot of changes since the last version number increase
(e.g the relocation of the RS485 communication code into the 'pibridge'
serdev driver module).
Also from now piControl is delivered as a debian package on its own, which
needs to have a proper version history in future.
So bump the module version to 2.0.0 as a starting point for this history.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>