piControl: set return value to 1 if firmware is already up to update
Currently 0 is returned by the ioctl PICONTROL_UPLOAD_FIRMWARE in both
cases, successful firmware upload and skipped upload if the firmware is
already up to date (and not force upload is specified).
Differentiate both cases by returning 1 instead of 0 in the latter case.

Furthermore make a full piControl reset also for this case, since it
simplifies the implementation and is not noticeably longer than calling
PiBridgeMaster_Continue() instead.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>