Return info for address 0 with KB_GET_DEVICE_INFO
KB_GET_DEVICE_INFO returns the info of a device except the RevPi. The
behaviour of this is inconsistent with KB_GET_DEVICE_INFO_LIST, which
returns the information of all devices, including the RevPi.

Fix this inconsistency by allowing address 0 for KB_GET_DEVICE_INFO to
return information about the RevPi and split the case block of
KB_GET_DEVICE_INFO into multiple functions for code reuse.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>